Functions separated in another cpp file
